Increase capacity variables to 16bit.

+++ examples/UPS/UPS.ino
@@ -26,8 +26,8 @@ uint16_t iRunTimeToEmpty = 0, iPrevRunTimeToEmpty = 0;
 uint16_t iManufacturerDate = 0; // initialized in setup function
 
 // Parameters for ACPI compliancy
-const byte iDesignCapacity = 100;
-byte iFullChargeCapacity = 100;
+const uint16_t iDesignCapacity = 100;
+uint16_t iFullChargeCapacity = 100;
 
 uint32_t iRemaining =0, iPrevRemaining=0;
 bool bCharging = false;
